feat: implement OOP approch
With the increment of the number of features, maintaining the main
script (FTIR-GUI.py) file become much difficult. Therfore based on OOP
approch, using tabview widget of customtkinter, two tabs are created
and those tabs contain main two features that are provided by this
tool. All UI elements and other functionalities are attached to
the corresponding tab and feature.